Renamed `Ethash` to `Blakimoto` since the most defining features of `Ethash` have been removed.
