+++ miner/blakimoto/readme.md
@@ -1,0 +1,6 @@
+### Blakimoto
+
+Blakimoto is the PoW mining algorithm. It is based on Ethash but without the memory hard 
+mining capabilities. Blakimoto is intended to be a pure PoW implementation using
+Blake2b-256 as the hashing function. Current difficulty algorithm is the same as 
+Ethash Frontier difficulty formula.
