+++ miner/blakimoto/sealer.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-package ethash
+package blakimoto
 
 import (
 	crand "crypto/rand"
@@ -31,37 +31,35 @@ import (
 
 // Seal implements consensus.Engine, attempting to find a nonce that satisfies
 // the block's difficulty requirements.
-func (ethash *Ethash) Seal(block *wire.Block, stop <-chan struct{}) (*wire.Block, error) {
+func (b *Blakimoto) Seal(block *wire.Block, stop <-chan struct{}) (*wire.Block, error) {
+
 	// If we're running a fake PoW, simply return a 0 nonce immediately
-	if ethash.config.PowMode == ModeFake || ethash.config.PowMode == ModeFullFake {
+	if b.config.PowMode == ModeTest {
 		header := block.GetHeader()
 		header.Nonce, header.MixHash = wire.EncodeNonce(0), util.Hash{}
 		block.SetHeader(header)
 
 		// delay for the specified time
-		time.Sleep(ethash.fakeDelay)
+		time.Sleep(b.fakeDelay)
 
 		return block, nil
 	}
-	// If we're running a shared PoW, delegate sealing to it
-	if ethash.shared != nil {
-		return ethash.shared.Seal(block, stop)
-	}
+
 	// Create a runner and the multiple search threads it directs
 	abort := make(chan struct{})
 	found := make(chan *wire.Block)
 
-	ethash.lock.Lock()
-	threads := ethash.threads
-	if ethash.rand == nil {
+	b.lock.Lock()
+	threads := b.threads
+	if b.rand == nil {
 		seed, err := crand.Int(crand.Reader, big.NewInt(math.MaxInt64))
 		if err != nil {
-			ethash.lock.Unlock()
+			b.lock.Unlock()
 			return nil, err
 		}
-		ethash.rand = rand.New(rand.NewSource(seed.Int64()))
+		b.rand = rand.New(rand.NewSource(seed.Int64()))
 	}
-	ethash.lock.Unlock()
+	b.lock.Unlock()
 	if threads == 0 {
 		threads = runtime.NumCPU()
 	}
@@ -74,8 +72,8 @@ func (ethash *Ethash) Seal(block *wire.Block, stop <-chan struct{}) (*wire.Block
 		pend.Add(1)
 		go func(id int, nonce uint64) {
 			defer pend.Done()
-			ethash.mine(block, id, nonce, abort, found)
-		}(i, uint64(ethash.rand.Int63()))
+			b.mine(block, id, nonce, abort, found)
+		}(i, uint64(b.rand.Int63()))
 	}
 	// Wait until sealing is terminated or a nonce is found
 	var result *wire.Block
@@ -86,11 +84,11 @@ func (ethash *Ethash) Seal(block *wire.Block, stop <-chan struct{}) (*wire.Block
 	case result = <-found:
 		// One of the threads found a block, abort all others
 		close(abort)
-	case <-ethash.update:
+	case <-b.update:
 		// Thread count was changed on user request, restart
 		close(abort)
 		pend.Wait()
-		return ethash.Seal(block, stop)
+		return b.Seal(block, stop)
 	}
 	// Wait for all miners to terminate and return the block
 	pend.Wait()
@@ -99,44 +97,42 @@ func (ethash *Ethash) Seal(block *wire.Block, stop <-chan struct{}) (*wire.Block
 
 // mine is the actual proof-of-work miner that searches for a nonce starting from
 // seed that results in correct final block difficulty.
-func (ethash *Ethash) mine(block *wire.Block, id int, seed uint64, abort chan struct{}, found chan *wire.Block) {
+func (b *Blakimoto) mine(block *wire.Block, id int, seed uint64, abort chan struct{}, found chan *wire.Block) {
 	// Extract some data from the header
 	var (
-		header  = block.GetHeader()
-		hash    = header.HashNoNonce().Bytes()
-		target  = new(big.Int).Div(maxUint256, header.Difficulty)
-		number  = header.Number
-		dataset = ethash.dataset(number)
+		header = block.GetHeader()
+		hash   = header.HashNoNonce().Bytes()
+		target = new(big.Int).Div(maxUint256, header.Difficulty)
 	)
 	// Start generating random nonces until we abort or find a good one
 	var (
 		attempts = int64(0)
 		nonce    = seed
 	)
-	log.Debug("Started ethash search for new nonces", "seed", seed, "id", id)
+	log.Debug("Started blakimoto search for new nonces", "seed", seed, "id", id)
 search:
 	for {
 		select {
 		case <-abort:
 			// Mining terminated, update stats and abort
 			log.Debug("Ethash nonce search aborted", "attempts", nonce-seed, "id", id)
-			ethash.hashrate.Mark(attempts)
+			b.hashrate.Mark(attempts)
 			break search
 
 		default:
 			// We don't have to update hash rate on every nonce, so update after after 2^X nonces
 			attempts++
 			if (attempts % (1 << 15)) == 0 {
-				ethash.hashrate.Mark(attempts)
+				b.hashrate.Mark(attempts)
 				attempts = 0
 			}
 			// Compute the PoW value of this nonce
-			digest, result := hashimotoFull(dataset.dataset, hash, nonce)
+			result := blakimoto(hash, nonce)
 			if new(big.Int).SetBytes(result).Cmp(target) <= 0 {
 				// Correct nonce found, create a new header with it
 				header = wire.CopyHeader(header)
 				header.Nonce = wire.EncodeNonce(nonce)
-				header.MixHash = util.BytesToHash(digest)
+				// header.MixHash = util.BytesToHash(digest)
 
 				// Seal and return a block (if still needed)
 				select {
@@ -150,7 +146,4 @@ search:
 			nonce++
 		}
 	}
-	// Datasets are unmapped in a finalizer. Ensure that the dataset stays live
-	// during sealing so it's not unmapped while being read.
-	runtime.KeepAlive(dataset)
 }
